refactor(Forms): Forms refactored and moved into the forms folder
