+++ frontend/src/app/admin/appointments/[id]/update/page.jsx
@@ -1,8 +1,7 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import { getAppointment } from "../../api/appointments";
-import { appointmentsRoute } from "@/app/utils/routes";
-import AppointmentForm from "../../AppointmentForm/AppointmentForm";
+import AppointmentForm from "../../../forms/AppointmentForm/AppointmentForm";
 import Layout from "@/app/components/base_layout/Layout";
 
 export default function UpdateAppointment({ params }) {
