+++ frontend/src/app/admin/forms/AppointmentDetail/AppointmentDetail.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { cancelAppointment, completeAppointment, getAppointment } from "../api/appointments";
+import { cancelAppointment, completeAppointment, getAppointment } from "../../appointments/api/appointments";
 import styles from "./Appointment-Detail.module.css";
 import { Status } from "@/app/components/form/status/Status";
 import Buttons from "@/app/components/form/model_buttons/Buttons";
