--- frontend/src/app/admin/services/ServiceForm/styles.module.css
@@ -1,117 +1,0 @@
-.fieldsContainer {
-    display: grid;
-    background-color: rgb(255, 255, 255);
-    box-shadow: 0px 2px 4px rgb(219, 219, 219);
-    box-sizing: border-box;
-    width: 100%;
-    border-radius: 10px;
-    margin-top: 1%;
-    margin-bottom: 1%;
-
-}
-
-.row {
-    padding: 1%;
-    box-sizing: content-box;
-    display: flex;
-}
-
-.row div {
-    margin-left: 1%;
-    width: 48%
-}
-
-.buttons {
-    display: flex;
-    gap: 2em;
-    justify-self: end;
-    margin-bottom: 1%;
-    margin-right: 1%;
-}
-
-.servicesContainer {
-    display:flex;
-    flex-wrap: wrap;
-    gap: 2em;
-    border-style: none;
-
-}
-
-.soloRow>label {
-    font-weight: 600;
-    font-size: 0.9rem;
-}
-
-.buttons button {
-    background-color: #0033cb;
-    color: white;
-    padding: 0.75em 3em;
-    border-radius: 6px;
-    font-weight: bold;
-    font-size: 0.9rem;
-}
-
-.buttons .cancelButton {
-    background-color: #4e4e4e;
-}
-
-
-.soloRow {
-    margin-left: 1%;
-    padding: 1%;
-    box-sizing: content-box;
-}
-
-.cancelBoton {
-    background-color: black;
-    color: white;
-    padding: 1%;
-}
-
-.fieldsConta {
-    display: block;
-}
-.fieldsTitle {
-    font-weight: 600;
-    font-size: 0.9rem;
-}
-
-.errorMessage {
-  color: #dc2626;
-  margin-bottom: 1rem;   
-}
-
-
-@media (width <=800px) {
-    .fieldsContainer {
-        min-width: fit-content;
-        display: grid;
-        margin: 0
-    }
-
-    .fieldsConta {
-        display: flex;
-        margin-bottom: 5%;
-        justify-content: center;
-    }
-
-    .buttons {        
-        gap: 1em;
-    }
-    
-    .row {
-        display: grid;
-        margin: 1%;
-        display: flexbox;
-    }
-
-    .row div {
-        width: 100%;
-    }
-
-    .soloRow {
-        margin: 0;
-        margin-left: 2%;
-        display: flexbox;
-    }
-}
