+++ frontend/src/app/admin/services/[id]/page.jsx
@@ -3,7 +3,7 @@ import Layout from "@/app/components/base_layout/Layout";
 import React, { useEffect, useState } from "react";
 import layout from "../../Main.module.css";
 import styles from "./styles.module.css";
-import warning from "../ServiceForm/styles.module.css"
+import warning from "../../forms/styles.module.css"
 import { getService } from "../api/services";
 import Buttons from "@/app/components/form/model_buttons/Buttons";
 import { Status } from "@/app/components/form/status/Status";
