+++ frontend/src/app/admin/services/[id]/styles.module.css
@@ -1,5 +1,4 @@
-.price
-{
+.price {
     margin-top: 2rem;
     display: flex;
     gap: 2rem;
