+++ frontend/src/app/admin/staff/[id]/update/page.jsx
@@ -1,11 +1,11 @@
 'use client';
 import CreateNewLayout from "@/app/components/base_layout/CreateNew/CreateNewLayout";
 import { staffRoute } from "@/app/utils/routes";
-import EmployeeForm from "../../EmployeeForm/EmployeeForm";
+import EmployeeForm from "../../../forms/EmployeeForm";
 import { getEmployee, updateEmployee } from "../../api/employees";
 import React, { useState, useEffect }  from "react";
 import Layout from "@/app/components/base_layout/Layout";
-import styles from '../../EmployeeForm/styles.module.css'
+import styles from '../../../forms/EmployeeFormStyles.module.css'
 
 export default function updateStaff({params}){
     const [employee, setEmployee] = useState(null);
